requirements/base.txt
Django==5.0.2
django-environ==0.11.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
drf-yasg==1.21.7
cerberus==1.3.5
